refactor: timeline・growth系コンポーネントからロジックを抽出
