+++ src/components/timeline/CommentSection.tsx
@@ -1,22 +1,7 @@
 "use client";
 
-import { useState, useEffect, useCallback, useRef } from "react";
 import { X, Send, Loader2, Trash2 } from "lucide-react";
-
-type CommentUser = {
-  id: string;
-  name: string;
-  avatar_url: string | null;
-};
-
-type Comment = {
-  id: string;
-  post_id: string;
-  user_id: string;
-  content: string;
-  created_at: string;
-  user: CommentUser;
-};
+import { useComments } from "./useComments";
 
 type CommentSectionProps = {
   postId: string;
@@ -33,87 +18,18 @@ export function CommentSection({
   currentUserId,
   onCommentCountChange,
 }: CommentSectionProps) {
-  const [comments, setComments] = useState<Comment[]>([]);
-  const [newComment, setNewComment] = useState("");
-  const [isLoading, setIsLoading] = useState(true);
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [deletingId, setDeletingId] = useState<string | null>(null);
-  const inputRef = useRef<HTMLInputElement>(null);
-  const commentsEndRef = useRef<HTMLDivElement>(null);
-
-  const fetchComments = useCallback(async () => {
-    setIsLoading(true);
-    try {
-      const response = await fetch(`/api/posts/${postId}/comments`);
-      if (response.ok) {
-        const data = await response.json();
-        setComments(data.comments);
-      }
-    } catch (error) {
-      console.error("Failed to fetch comments:", error);
-    } finally {
-      setIsLoading(false);
-    }
-  }, [postId]);
-
-  useEffect(() => {
-    if (isOpen) {
-      fetchComments();
-      setTimeout(() => inputRef.current?.focus(), 100);
-    }
-  }, [isOpen, fetchComments]);
-
-  useEffect(() => {
-    if (comments.length > 0) {
-      commentsEndRef.current?.scrollIntoView({ behavior: "smooth" });
-    }
-  }, [comments.length]);
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    if (!newComment.trim() || isSubmitting) return;
-
-    setIsSubmitting(true);
-    try {
-      const response = await fetch(`/api/posts/${postId}/comments`, {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ content: newComment.trim() }),
-      });
-
-      if (response.ok) {
-        const data = await response.json();
-        setComments((prev) => [...prev, data.comment]);
-        setNewComment("");
-        onCommentCountChange?.(postId, comments.length + 1);
-      }
-    } catch (error) {
-      console.error("Failed to submit comment:", error);
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
-
-  const handleDelete = async (commentId: string) => {
-    if (deletingId) return;
-
-    setDeletingId(commentId);
-    try {
-      const response = await fetch(
-        `/api/posts/${postId}/comments/${commentId}`,
-        { method: "DELETE" }
-      );
-
-      if (response.ok) {
-        setComments((prev) => prev.filter((c) => c.id !== commentId));
-        onCommentCountChange?.(postId, comments.length - 1);
-      }
-    } catch (error) {
-      console.error("Failed to delete comment:", error);
-    } finally {
-      setDeletingId(null);
-    }
-  };
+  const {
+    comments,
+    newComment,
+    setNewComment,
+    isLoading,
+    isSubmitting,
+    deletingId,
+    inputRef,
+    commentsEndRef,
+    handleSubmit,
+    handleDelete,
+  } = useComments({ postId, isOpen, onCommentCountChange });
 
   const formatDate = (dateString: string) => {
     const date = new Date(dateString);
